refactoring: tokenが無効だと、ログイン画面にリダイレクトさせる処理をまとめる
