+++ front/api/client/icon_get.tsx
@@ -1,6 +1,6 @@
 import { GenerateImageUrl } from "../../component/icon"
 import { AxiosResponse } from 'axios';
-import { privateClient } from '../../libs/axios'
+import { privateClient } from '../../libs/client/axios'
 
 export const UserIconGet = async (icon: string) => {
     const url: string = 'v1/icon/' + icon
